% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2014}
\alias{accident_2014}
\title{Fatal injuries suffered in motor vehicle traffic crashes in the US, 2014}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30056 rows and 50 columns.
}
\source{
NHTSA - Fatality Analysis Reporting System (FARS) \url{https://www.nhtsa.gov/node/97996/251}
}
\usage{
accident_2014
}
\description{
Fatal injuries suffered in motor vehicle traffic crashes in the US, 2014
}
\examples{
\dontrun{
 accident_2013
}
}
\keyword{datasets}
