% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accts_recv}
\alias{accts_recv}
\title{accts_recv}
\format{
A data frame with 2466 rows and 12 variables:
\describe{
  \item{\code{countryCode}}{integer:  country identifier}
  \item{\code{customerID}}{character:  customer identifier}
  \item{\code{PaperlessDate}}{character:  date the company began paperless billing}
  \item{\code{invoiceNumber}}{character:  invoice identifier}
  \item{\code{InvoiceDate}}{character:  date of invoice}
  \item{\code{DueDate}}{character:  day payment due}
  \item{\code{InvoiceAmount}}{double:  amount of invoice, in $}
  \item{\code{Disputed}}{character:  Yes if invoice was disputed; No otherwise.}
  \item{\code{SettledDate}}{character:  Date that invoice was paid or settled.}
  \item{\code{PaperlessBill}}{character:  Paper if a hard copy of the invoice was sent; Electronic otherwise}
  \item{\code{DaysToSettle}}{integer:  number of days required to settle the invoice}
  \item{\code{DaysLate}}{integer:  number of days late the invoice was} 
}
}
\source{
\url{https://www.ibm.com/communities/analytics/watson-analytics-blog/guide-to-sample-datasets/}
}
\usage{
accts_recv
}
\description{
simulated data from IBM Watson about "Understand "the factors of successful collection efforts. You can Predict which customers will pay fastest and recover more money and improve collections efficiency."
}
\keyword{datasets}
