% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AdsManager}
\alias{AdsManager}
\title{AdsManager}
\format{A data frame with 2667 rows and 6 variables:
\describe{
  \item{\code{platform}}{character did viewer see ad on Facebook or Instagram}
  \item{\code{age_group}}{character age group of viewer}
  \item{\code{device_type}}{character device type viewer used to see ad}
  \item{\code{clicked}}{character yes if viewer clicked on the ad to navigate to website}
  \item{\code{pageviews}}{double number of pages viewed after clicking}
  \item{\code{amt_spent}}{double amount spent on website after clicking} 
}}
\source{
"made-up by Stephanie Jernigan"
}
\usage{
AdsManager
}
\description{
Fields from Google AdsManager
}
\keyword{datasets}
