% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adultdata}
\alias{adultdata}
\title{adultdata}
\format{
A data frame with 32561 rows and 15 variables:
\describe{
  \item{\code{age}}{double:  age in years}
  \item{\code{workclass}}{character categories: Private, Self-emp-not-inc, Self-emp-inc, Federal-gov, Local-gov, State-gov, Without-pay, Never-worked}
  \item{\code{fnlwgt}}{double final weight index determined by: the estimated population 16+ for each state, origin, race, age and sex. People with similar demographic characteristics should have similar weights. see https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.names for a more detailed description}
  \item{\code{education}}{character categories: Bachelors, Some-college, 11th, HS-grad, Prof-school, Assoc-acdm, Assoc-voc, 9th, 7th-8th, 12th, Masters, 1st-4th, 10th, Doctorate, 5th-6th, Preschool}
  \item{\code{education_num}}{double number of education years}
  \item{\code{marital_status}}{character categories: Married-civ-spouse, Divorced, Never-married, Separated, Widowed, Married-spouse-absent, Married-AF-spouse}
  \item{\code{occupation}}{character categories: Tech-support, Craft-repair, Other-service, Sales, Exec-managerial, Prof-specialty, Handlers-cleaners, Machine-op-inspct, Adm-clerical, Farming-fishing, Transport-moving, Priv-house-serv, Protective-serv, Armed-Forces}
  \item{\code{relationship}}{character categories: relationship: Wife, Own-child, Husband, Not-in-family, Other-relative, Unmarried}
  \item{\code{race}}{character categories: White, Asian-Pac-Islander, Amer-Indian-Eskimo, Other, Black}
  \item{\code{sex}}{character categories: Female, Male}
  \item{\code{capital_gain}}{double capital gain from the year}
  \item{\code{capital_loss}}{double capital loss from the year}
  \item{\code{hours_per_week}}{double number of hours worked per week}
  \item{\code{native_country}}{character categories: United-States, Cambodia, England, Puerto-Rico, Canada, Germany, Outlying-US(Guam-USVI-etc), India, Japan, Greece, South, China, Cuba, Iran, Honduras, Philippines, Italy, Poland, Jamaica, Vietnam, Mexico, Portugal, Ireland, France, Dominican-Republic, Laos, Ecuador, Taiwan, Haiti, Columbia, Hungary, Guatemala, Nicaragua, Scotland, Thailand, Yugoslavia, El-Salvador, Trinadad&Tobago, Peru, Hong, Holand-Netherlands}
  \item{\code{morethan50k}}{character categories: >50K (annual income greater than $50K), <=50K (annual income less than $50K)} 
}
}
\source{
\url{http://www.census.gov/ftp/pub/DES/www/welcome.html}
}
\usage{
adultdata
}
\description{
Data extracted from the 1994 census bureau database
}
\keyword{datasets}
