% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AirlineData}
\alias{AirlineData}
\title{AirlineData}
\format{A data frame with 50 rows and 4 variables:
\describe{
  \item{\code{airline}}{character:  airline}
  \item{\code{status}}{character:  was flight delayed or not}
  \item{\code{minutes}}{integer:  minutes of delay}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
AirlineData
}
\description{
sample of flights from two airlines, Jet Blue and Southwest
}
\keyword{datasets}
