% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AttendGrad}
\alias{AttendGrad}
\title{AttendGrad}
\format{A data frame with 97 rows and 1 variables:
\describe{
  \item{\code{attend_grad}}{character Yes, No, or Not sure} 
}}
\source{
{past survey in Business Statistics}
}
\usage{
AttendGrad
}
\description{
data from past class survey about whether or not students plan to attend grad school
}
\keyword{datasets}
