% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atus_summary}
\alias{atus_summary}
\title{atus_summary}
\format{
A data frame with 130150 rows and 23 variables:
\describe{
  \item{\code{Education Level}}{character COLUMN_DESCRIPTION}
  \item{\code{Age}}{integer COLUMN_DESCRIPTION}
  \item{\code{Age Range}}{character COLUMN_DESCRIPTION}
  \item{\code{Employment Status}}{character COLUMN_DESCRIPTION}
  \item{\code{Gender}}{character COLUMN_DESCRIPTION}
  \item{\code{Children}}{integer COLUMN_DESCRIPTION}
  \item{\code{Weekly Earnings}}{integer COLUMN_DESCRIPTION}
  \item{\code{Year}}{integer COLUMN_DESCRIPTION}
  \item{\code{Weekly Hours Worked}}{integer COLUMN_DESCRIPTION}
  \item{\code{Sleeping}}{integer COLUMN_DESCRIPTION}
  \item{\code{Grooming}}{integer COLUMN_DESCRIPTION}
  \item{\code{Housework}}{integer COLUMN_DESCRIPTION}
  \item{\code{Food & Drink Prep}}{integer COLUMN_DESCRIPTION}
  \item{\code{Caring for Children}}{integer COLUMN_DESCRIPTION}
  \item{\code{Playing with Children}}{integer COLUMN_DESCRIPTION}
  \item{\code{Job Searching}}{integer COLUMN_DESCRIPTION}
  \item{\code{Shopping}}{integer COLUMN_DESCRIPTION}
  \item{\code{Eating and Drinking}}{integer COLUMN_DESCRIPTION}
  \item{\code{Socializing & Relaxing}}{integer COLUMN_DESCRIPTION}
  \item{\code{Television}}{integer COLUMN_DESCRIPTION}
  \item{\code{Golfing}}{integer COLUMN_DESCRIPTION}
  \item{\code{Running}}{integer COLUMN_DESCRIPTION}
  \item{\code{Volunteering}}{integer COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://www.ibm.com/communities/analytics/watson-analytics-blog/guide-to-sample-datasets/}
}
\usage{
atus_summary
}
\description{
Summary of results of US gov't American Time Use Survey
}
\keyword{datasets}
