% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bikeshare}
\alias{bikeshare}
\title{bikeshare}
\format{
A data frame with 17379 rows and 12 variables:
\describe{
  \item{\code{Instant}}{integer:  time slice identifier}
  \item{\code{Riders}}{integer:  number of riders during at this time}
  \item{\code{Season}}{integer:  1 = winter, 2 = spring, 3 = summer, 4 = fall}
  \item{\code{Month}}{integer:  month}
  \item{\code{Hour}}{integer:  hour}
  \item{\code{Holiday}}{integer:  1 if holiday; 0 otherwise}
  \item{\code{Weekday}}{integer:  0 through 6 representing the day of the week, Monday to Sunday}
  \item{\code{Workday}}{integer:  1 if workday; 0 otherwise}
  \item{\code{Weather}}{integer:  general category of weather.  Exact definition unknown.}
  \item{\code{Temperature}}{integer:  temperature in degrees F}
  \item{\code{Humidity}}{double:  relative humidity measure}
  \item{\code{Wind}}{double:  an unknown measure of wind speed} 
}
}
\source{
\url{https://www.ibm.com/communities/analytics/watson-analytics-blog/guide-to-sample-datasets/}
}
\usage{
bikeshare
}
\description{
hourly information about a public bike share program in Washington, D.C.
}
\keyword{datasets}
