% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bk}
\alias{bk}
\title{bk}
\format{A data frame with 32 rows and 9 variables:
\describe{
  \item{\code{serving}}{double:  serving size, in grams}
  \item{\code{calories}}{double:  calories per serving}
  \item{\code{carbs}}{double:  grams of carbohydrates}
  \item{\code{meat}}{double:  1 if item contains meat; 0 otherwise}
  \item{\code{total_fat}}{double:  grams of fat per serving}
  \item{\code{cholesterol}}{double:  milligrams of cholesterol per serving}
  \item{\code{sodium}}{double milligrams of sodium per serving}
  \item{\code{sugars}}{double:  grams of sugar per serving}
  \item{\code{protein}}{double:  grams of protein per serving} 
}}
\source{
{Sharpe, De Veaux, Velleman:  Business Statistics, A First Course, 2e, online chapters}
}
\usage{
bk
}
\description{
nutrition items for foods at Burger King
}
\keyword{datasets}
