% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bluebook}
\alias{bluebook}
\title{bluebook}
\format{A data frame with 804 rows and 3 variables:
\describe{
  \item{\code{Price}}{double:  sales price}
  \item{\code{Mileage}}{double: miles on odometer at time of sale}
  \item{\code{Make}}{character:  car manufacturer} 
}}
\source{
\url{Jaggia and Kelly textbook}
}
\usage{
bluebook
}
\description{
sample of used car attributes and sales prices
}
\keyword{datasets}
