% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burgerking}
\alias{burgerking}
\title{burgerking}
\format{A data frame with 32 rows and 10 variables:
\describe{
  \item{\code{item}}{character:  menu item name}
  \item{\code{Serving}}{double:  serving size in grams}
  \item{\code{Calories}}{double:  number of calories per serving}
  \item{\code{Carbs}}{double:  grams of carbohydrates per serving}
  \item{\code{Meat}}{double:  1 if contains meat; 0 otherwise}
  \item{\code{Total Fat}}{double:  grams of fat per serving}
  \item{\code{Cholesterol}}{double:  milligrams of cholesterol per serving}
  \item{\code{Sodium}}{double:  milligrams of sodium per serving}
  \item{\code{Sugars}}{double:  grams of sugar per serving}
  \item{\code{Protein}}{double:  grams of protein per serving} 
}}
\source{
\url{Sharpe, DeVeaux, Velleman electronic chapters}
}
\usage{
burgerking
}
\description{
nutrition information for Burger King menu items
}
\keyword{datasets}
