% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardata}
\alias{cardata}
\title{cardata}
\format{
A data frame with 804 rows and 3 variables:
\describe{
  \item{\code{Price}}{double: price of used car in US$}
  \item{\code{Mileage}}{double:  miles on odometer at time of sale}
  \item{\code{Make}}{character:  brand of car} 
}
}
\source{
{}
}
\usage{
cardata
}
\description{
data about used cars from the Kelley Blue book
}
\keyword{datasets}
