% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarInsData}
\alias{CarInsData}
\title{CarInsData}
\format{
A data frame with 500 rows and 31 variables:
\describe{
  \item{\code{Obs}}{double }
  \item{\code{ID}}{double }
  \item{\code{Crash}}{double 0/1 indicator for crash where 1 indicates a crash}
  \item{\code{CrashCost}}{double cost to insurance company of crash}
  \item{\code{DriverAge}}{double age of driver}
  \item{\code{LevelEducation}}{character level education of driver}
  \item{\code{YearsJob}}{double years on current job}
  \item{\code{Income}}{double income in US dollars}
  \item{\code{SexofDriver}}{character sex, male or female}
  \item{\code{TypeofJob}}{character job type}
  \item{\code{TimeCommute}}{double in minutes, length of regular work commute}
  \item{\code{PointsonRecord}}{double number points on driving record}
  \item{\code{LicenseRevoked}}{character 0/1 indicator, 1 indicates has had license revoked}
  \item{\code{ValueofHome}}{double in US dollars, value of home}
  \item{\code{MaritalStatus}}{character indicator marital status}
  \item{\code{SingleParent}}{character indicator single parent status}
  \item{\code{NumberKidsHome}}{double }
  \item{\code{NumberKidsDrive}}{double }
  \item{\code{UrbanorRural}}{character where driver typically lives & drives}
  \item{\code{AgeofCar}}{double }
  \item{\code{TypeofCar}}{character }
  \item{\code{TypeCarUse}}{character how vehicle is typically used}
  \item{\code{BlueBookValue}}{double replacement value from Blue Book}
  \item{\code{RedCar}}{character indicator for red cars}
  \item{\code{TimeasCustomer}}{double number years driver has been a customer}
  \item{\code{NumberClaims5yr}}{double number of claims in last 5 years}
  \item{\code{Total5YrPayout}}{double US dollar value payouts of claims in last 5 years}
  \item{\code{NotWorking}}{character indicator working or not}
  \item{\code{NewCar}}{character indicator, new car or not}
  \item{\code{GoodDriver}}{character indicator, good driver or not, based on points}
  \item{\code{OwnHome}}{character indicator, own or rent, based on home value} 
}
}
\usage{
CarInsData
}
\description{
dataset that includes various attributes about drivers, cars, and car crashes
}
\keyword{datasets}
