% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoffeeSaleData}
\alias{CoffeeSaleData}
\title{CoffeeSaleData}
\format{A data frame with 50 rows and 1 variables:
\describe{
  \item{\code{sales}}{double:  purchase amount, in dollars} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
CoffeeSaleData
}
\description{
per-person purchase amount of a sample of 50 coffee shop customers
}
\keyword{datasets}
