% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoffeeTypeData}
\alias{CoffeeTypeData}
\title{CoffeeTypeData}
\format{
A data frame with 50 rows and 1 variables:
\describe{
  \item{\code{type}}{character: purchaser type} 
}
}
\source{
{}
}
\usage{
CoffeeTypeData
}
\description{
Customer type of coffee purchases
}
\keyword{datasets}
