% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CollegeRet}
\alias{CollegeRet}
\title{CollegeRet}
\format{A data frame with 603 rows and 4 variables:
\describe{
  \item{\code{currentyear}}{character:  class of student}
  \item{\code{status}}{character:  are they returning to school or not}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
CollegeRet
}
\description{
A sample of students and whether or not they are returning to school
}
\keyword{datasets}
