% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CollegeRetData}
\alias{CollegeRetData}
\title{CollegeRetData}
\format{
A data frame with 603 rows and 4 variables:
\describe{
  \item{\code{currentyear}}{character: student's year in college}
  \item{\code{status}}{character: student's returning status for next college year}
}
}
\source{
{}
}
\usage{
CollegeRetData
}
\description{
Retention data for college students
}
\keyword{datasets}
