% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complaints}
\alias{complaints}
\title{complaints}
\format{A data frame with 31 rows and 2 variables:
\describe{
  \item{\code{complaints_before}}{double:  average complaints per week before training}
  \item{\code{complaints_after}}{double:  average complaints per week after training} 
}}
\source{
{}
}
\usage{
complaints
}
\description{
complaints recorded about customer service agents before training and after training
}
\keyword{datasets}
