% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConcChoice}
\alias{ConcChoice}
\title{ConcChoice}
\format{A data frame with 49 rows and 1 variables:
\describe{
  \item{\code{conc_choice}}{character:  The name of the concentration they chose.} 
}}
\source{
{past survey of Business Statistics students}
}
\usage{
ConcChoice
}
\description{
Business Statistics students' choices of concentrations if they had to decide today.
}
\keyword{datasets}
