% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CreditCardData}
\alias{CreditCardData}
\title{CreditCardData}
\format{A data frame with 401 rows and 11 variables:
\describe{
  \item{\code{id}}{integer: customer ID}
  \item{\code{income}}{double:  Income, in $1000s}
  \item{\code{limit}}{integer: Credit limit, in $}
  \item{\code{rating}}{integer:  Credit rating score from 0 to 1000 possible}
  \item{\code{cards}}{integer:  total number of credit cards customer has}
  \item{\code{age}}{integer:  Customer age in years}
  \item{\code{education}}{integer:  years of education (12 = high school)}
  \item{\code{gender}}{character:  Customer gender}
  \item{\code{student}}{character:  1 if student; 0 otherwise}
  \item{\code{married}}{character:  1 if married; 0 otherwise}
  \item{\code{balance}}{integer:  Average monthly balance, in $} 
}}
\source{
{adjusted from a sample of 'credit' dataset in 'ISLR' package.}
}
\usage{
CreditCardData
}
\description{
information about 401 credit card customers.
}
\keyword{datasets}
