% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creditcards}
\alias{creditcards}
\title{creditcards}
\format{A data frame with 400 rows and 11 variables:
\describe{
  \item{\code{ID}}{integer: customer ID}
  \item{\code{Income}}{double:  Income, in $1000s}
  \item{\code{Limit}}{integer: Credit limit, in $}
  \item{\code{Rating}}{integer:  Credit rating score from 0 to 1000 possible.}
  \item{\code{Cards}}{integer:  total number of credit cards customer has}
  \item{\code{Age}}{integer:  Customer age in years}
  \item{\code{Education}}{integer:  years of education (12 = high school)}
  \item{\code{Gender}}{character:  Customer gender}
  \item{\code{Student}}{character:  1 if student; 0 otherwise}
  \item{\code{Married}}{character:  1 if married; 0 otherwise}
  \item{\code{Balance}}{integer:  Average monthly balance, in $} 
}}
\source{
{adjusted from a sample of 'credit' dataset in 'ISLR' package.}
}
\usage{
creditcards
}
\description{
Credit card customer data
}
\keyword{datasets}
