% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decindep}
\alias{decindep}
\title{decindep}
\format{A data frame with 1323 rows and 3 variables:
\describe{
  \item{\code{Word}}{character:  a word in the Declaration of Independence}
  \item{\code{Length}}{integer:  number of letters in the word}
  \item{\code{Has_an_e}}{integer:  1 if the word contains an e; 0 otherwise} 
}}
\source{
{uncredited version of the Declaration of Independence, edited to remove dates, signatures, etc.}
}
\usage{
decindep
}
\description{
words in the Declaration of Independence
}
\keyword{datasets}
