% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fns.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{A function showing descriptive statistics}
\usage{
descriptive_stats(num_list)
}
\arguments{
\item{num_list}{}
}
\value{
A list of descriptive info
}
\description{
A function showing descriptive statistics
}
