% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame representing the graph with columns: v1 (from node), v2 (to node), w (weight)}

\item{init_node}{The starting node}
}
\value{
A vector of shortest distances from the starting node to all other nodes
}
\description{
This function finds the shortest path from a starting node to all other nodes
in a graph using Dijkstra's algorithm.
}
\examples{
wiki_graph <- data.frame(v1 = c(1, 1, 1), v2 = c(2, 3, 6), w = c(7, 9, 14))
dijkstra(wiki_graph, 1)
}
