% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distfromhome}
\alias{distfromhome}
\title{distfromhome}
\format{
A data frame with 375 rows and 2 variables:
\describe{
  \item{\code{School}}{character:  School identifier}
  \item{\code{Distance}}{double:  distance from home in miles} 
}
}
\source{
{made up for an in-class exercise.  BC data comes from a class survey.}
}
\usage{
distfromhome
}
\description{
Simulated data set with distance from home (miles) of students from 4 business programs
}
\keyword{datasets}
