% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distfromhome_bc}
\alias{distfromhome_bc}
\title{distfromhome_bc}
\format{A data frame with 97 rows and 2 variables:
\describe{
  \item{\code{School}}{character:  school identifier (only BC)}
  \item{\code{Distance}}{double:  distance from home in miles} 
}}
\source{
{from previous class survey}
}
\usage{
distfromhome_bc
}
\description{
dataset from previous class about distance students travel to attend BC.
}
\keyword{datasets}
