% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EdLevel}
\alias{EdLevel}
\title{EdLevel}
\format{
A data frame with 510 rows and 1 variables:
\describe{
  \item{\code{edlevel}}{character:  highest education level attained} 
}
}
\usage{
EdLevel
}
\description{
made-up data set for HW Ch. 4
}
\keyword{datasets}
