% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmailAccessData}
\alias{EmailAccessData}
\title{EmailAccessData}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{\code{id}}{integer: identifier}
  \item{\code{emailv1}}{double: Version 1 download speed in seconds}
  \item{\code{emailv2}}{double: Version 2 download speed in seconds} 
}}
\source{
{}
}
\usage{
EmailAccessData
}
\description{
Download speeds of two email systems
}
\keyword{datasets}
