% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmployeeAtt}
\alias{EmployeeAtt}
\title{EmployeeAtt}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1470 rows and 35 columns.}
\source{
\url{https://www.kaggle.com/pavansubhasht/ibm-hr-analytics-attrition-dataset/metadata}
}
\usage{
EmployeeAtt
}
\description{
simulated data on employee attrition created by IBM
}
\keyword{datasets}
