% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
The GCD of `a` and `b`
}
\description{
This function calculates the greatest common divisor (GCD) of two integers.
}
\examples{
euclidean(123612, 13892347912)  # Returns 4
euclidean(100, 1000)            # Returns 100
}
