% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evals}
\alias{evals}
\title{evals}
\format{A data frame with 463 rows and 21 variables:
\describe{
  \item{\code{score}}{double Average professor evaluation score}
  \item{\code{rank}}{integer Rank of professor:  teaching, tenure track, tenured.}
  \item{\code{ethnicity}}{integer Ethnicity of professor: not minority, minority}
  \item{\code{gender}}{integer Gender of professor: female, male.}
  \item{\code{language}}{integer Language of school where professor received education: English or non-English.}
  \item{\code{age}}{integer Language of school where professor received education: English or non-English.}
  \item{\code{cls_perc_eval}}{double Percent of students in class who completed evaluation.}
  \item{\code{cls_did_eval}}{integer Number of students in class who completed evaluation.}
  \item{\code{cls_students}}{integer Total number of students in class.}
  \item{\code{cls_level}}{integer Class level: lower, upper.}
  \item{\code{cls_profs}}{integer Number of professors teaching sections in course in sample: single, multiple.}
  \item{\code{cls_credits}}{integer Number of credits of class: one credit (lab, PE, etc.), multi credit.}
  \item{\code{bty_f1lower}}{integer Beauty rating of professor from lower level female: (1) lowest - (10) highest.}
  \item{\code{bty_f1upper}}{integer Beauty rating of professor from upper level female: (1) lowest - (10) highest.}
  \item{\code{bty_f2upper}}{integer Beauty rating of professor from second level female: (1) lowest - (10) highest.}
  \item{\code{bty_m1lower}}{integer Beauty rating of professor from lower level male: (1) lowest - (10) highest.}
  \item{\code{bty_m1upper}}{integer Beauty rating of professor from upper level male: (1) lowest - (10) highest.}
  \item{\code{bty_m2upper}}{integer Beauty rating of professor from second upper level male: (1) lowest - (10) highest.}
  \item{\code{bty_avg}}{double Average beauty rating of professor.}
  \item{\code{pic_outfit}}{integer Outfit of professor in picture: not formal, formal.}
  \item{\code{pic_color}}{integer Color of professor’s picture: color, black & white.} 
}}
\source{
\url{https://www.openintro.org/stat/data/evals.php}
}
\usage{
evals
}
\description{
The data are gathered from end of semester student evaluations for a large sample of professors from the University of Texas at Austin. In addition, six students rate the professors' physical appearance. The result is a data frame where each row contains a different course and each column has information on either the course or the professor. For more information, please see Çetinkaya-Rundel M, Morgan KL, Stangl D. 2013 (coming soon). Looking Good on Course Evaluations. CHANCE 26(2).
}
\keyword{datasets}
