% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{extracts complete zip and returns a filtered list of the content}
\usage{
extractData(zipFileName, workingfolder, NameFilter = ".*")
}
\arguments{
\item{zipFileName}{}

\item{workingfolder}{}

\item{NameFilter}{}
}
\value{

}
\description{
extracts complete zip and returns a filtered list of the content
}
