% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state identifier.}

\item{year}{An integer of the year to reference for mapping.}
}
\value{
A map plot showing accidents in the state within the year.
}
\description{
This function takes the supplied state number and year and plots the  
locations of accidents on a map outlining the specified state.
}
\details{
Note the maps library must be loaded in the session else it may fail.

If an invalid state number is supplied it will error saying this.

If there are no accidents to plot, it will return null.
}
\examples{
\dontrun{
map <- fars_map_state(state.num = 10, year = 2017)
map <- fars_map_state(10, 2017)
}
}
