% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A wide dataframe of the number of observations by month and year.
}
\description{
This function takes a vector of years and reads data from derived filenames 
then summarizes the number of observations for each year grouped by month. 
It transforms the data to wide format and returns this as a data frame.
}
\examples{
\dontrun{
df <- fars_summarize_years(years = c(2010:2017))
df <- fars_summarize_years(c(2010:2017))
}
}
