% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FilmAges}
\alias{FilmAges}
\title{FilmAges}
\format{
A data frame with 468 rows and 1 variables:
\describe{
  \item{\code{ages}}{character:  age groups who attended movie} 
}
}
\usage{
FilmAges
}
\description{
made-up data set for HW Ch. 4
}
\keyword{datasets}
