% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foodads}
\alias{foodads}
\title{foodads}
\format{A data frame with 118 rows and 2 variables:
\describe{
  \item{\code{AdType}}{character:  type of ad child watched, either food or nonfood}
  \item{\code{GramsEaten}}{double:  grams of Goldfish crackers child ate while watching} 
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2743554/}
}
\usage{
foodads
}
\description{
simulated data about amount of Goldfish crackers eaten by two different groups of children:  those who watched a TV program with food ads and those who watched the TV program with non-food ads.
}
\keyword{datasets}
