% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathHelpers.R
\name{getFilenameWithoutExtension}
\alias{getFilenameWithoutExtension}
\title{removes path and extension}
\usage{
getFilenameWithoutExtension(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
removes path and extension
}
