% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GoogleAnalyticsData}
\alias{GoogleAnalyticsData}
\title{GoogleAnalyticsData}
\format{A data frame with 5732 rows and 7 variables:
\describe{
  \item{\code{date}}{double date in Y-m-d format}
  \item{\code{channelGrouping}}{character how site viewer found site}
  \item{\code{deviceCategory}}{character type of device used to view site}
  \item{\code{sessions}}{integer total sessions:  "a group of user interactions with your website that take place within a given time frame"}
  \item{\code{pageviews}}{integer total pages viewed by those viewers}
  \item{\code{entrances}}{integer number of entrances, see https://support.google.com/analytics/answer/2956047?hl=en}
  \item{\code{bounces}}{integer viewers who leave site after viewing exactly one page}
}}
\source{
\url{http://www.dartistics.com/data/gadata_example_2.csv}
}
\usage{
GoogleAnalyticsData
}
\description{
Sample of Google Analytics data for a website over several months.  Possibly made up.
}
\keyword{datasets}
