% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HeartData}
\alias{HeartData}
\title{HeartData}
\format{A data frame with 303 rows and 14 variables:
\describe{
  \item{\code{age}}{double COLUMN_DESCRIPTION}
  \item{\code{sex}}{double COLUMN_DESCRIPTION}
  \item{\code{cp}}{double COLUMN_DESCRIPTION}
  \item{\code{trestbps}}{double COLUMN_DESCRIPTION}
  \item{\code{chol}}{double COLUMN_DESCRIPTION}
  \item{\code{fbs}}{double COLUMN_DESCRIPTION}
  \item{\code{restecg}}{double COLUMN_DESCRIPTION}
  \item{\code{thalach}}{double COLUMN_DESCRIPTION}
  \item{\code{exang}}{double COLUMN_DESCRIPTION}
  \item{\code{oldpeak}}{double COLUMN_DESCRIPTION}
  \item{\code{slope}}{double COLUMN_DESCRIPTION}
  \item{\code{ca}}{double COLUMN_DESCRIPTION}
  \item{\code{thal}}{double COLUMN_DESCRIPTION}
  \item{\code{target}}{double COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
HeartData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
