% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HouseSalesData}
\alias{HouseSalesData}
\title{HouseSalesData}
\format{A data frame with 1460 rows and 80 variables:
\describe{
  \item{\code{Id}}{double COLUMN_DESCRIPTION}
  \item{\code{MsZoning}}{character COLUMN_DESCRIPTION}
  \item{\code{LotFrontage}}{double COLUMN_DESCRIPTION}
  \item{\code{LotArea}}{double COLUMN_DESCRIPTION}
  \item{\code{Street}}{character COLUMN_DESCRIPTION}
  \item{\code{Alley}}{character COLUMN_DESCRIPTION}
  \item{\code{LotShape}}{character COLUMN_DESCRIPTION}
  \item{\code{LandContour}}{character COLUMN_DESCRIPTION}
  \item{\code{Utilities}}{character COLUMN_DESCRIPTION}
  \item{\code{LotConfig}}{character COLUMN_DESCRIPTION}
  \item{\code{LandSlope}}{character COLUMN_DESCRIPTION}
  \item{\code{Neighborhood}}{character COLUMN_DESCRIPTION}
  \item{\code{Condition1}}{character COLUMN_DESCRIPTION}
  \item{\code{Condition2}}{character COLUMN_DESCRIPTION}
  \item{\code{BldgType}}{character COLUMN_DESCRIPTION}
  \item{\code{HouseStyle}}{character COLUMN_DESCRIPTION}
  \item{\code{OverallQual}}{double COLUMN_DESCRIPTION}
  \item{\code{OverallCond}}{double COLUMN_DESCRIPTION}
  \item{\code{YearBuilt}}{double COLUMN_DESCRIPTION}
  \item{\code{YearRemodAdd}}{double COLUMN_DESCRIPTION}
  \item{\code{RoofStyle}}{character COLUMN_DESCRIPTION}
  \item{\code{RoofMatl}}{character COLUMN_DESCRIPTION}
  \item{\code{Exterior1St}}{character COLUMN_DESCRIPTION}
  \item{\code{Exterior2Nd}}{character COLUMN_DESCRIPTION}
  \item{\code{MasVnrType}}{character COLUMN_DESCRIPTION}
  \item{\code{MasVnrArea}}{double COLUMN_DESCRIPTION}
  \item{\code{ExterQual}}{character COLUMN_DESCRIPTION}
  \item{\code{ExterCond}}{character COLUMN_DESCRIPTION}
  \item{\code{Foundation}}{character COLUMN_DESCRIPTION}
  \item{\code{BsmtQual}}{character COLUMN_DESCRIPTION}
  \item{\code{BsmtCond}}{character COLUMN_DESCRIPTION}
  \item{\code{BsmtExposure}}{character COLUMN_DESCRIPTION}
  \item{\code{BsmtFinType1}}{character COLUMN_DESCRIPTION}
  \item{\code{BsmtFinSf1}}{double COLUMN_DESCRIPTION}
  \item{\code{BsmtFinType2}}{character COLUMN_DESCRIPTION}
  \item{\code{BsmtFinSf2}}{double COLUMN_DESCRIPTION}
  \item{\code{BsmtUnfSf}}{double COLUMN_DESCRIPTION}
  \item{\code{TotalBsmtSf}}{double COLUMN_DESCRIPTION}
  \item{\code{Heating}}{character COLUMN_DESCRIPTION}
  \item{\code{HeatingQc}}{character COLUMN_DESCRIPTION}
  \item{\code{CentralAir}}{character COLUMN_DESCRIPTION}
  \item{\code{Electrical}}{character COLUMN_DESCRIPTION}
  \item{\code{X1StFlrSf}}{double COLUMN_DESCRIPTION}
  \item{\code{X2NdFlrSf}}{double COLUMN_DESCRIPTION}
  \item{\code{LowQualFinSf}}{double COLUMN_DESCRIPTION}
  \item{\code{GrLivArea}}{double COLUMN_DESCRIPTION}
  \item{\code{BsmtFullBath}}{double COLUMN_DESCRIPTION}
  \item{\code{BsmtHalfBath}}{double COLUMN_DESCRIPTION}
  \item{\code{FullBath}}{double COLUMN_DESCRIPTION}
  \item{\code{HalfBath}}{double COLUMN_DESCRIPTION}
  \item{\code{BedroomAbvGr}}{double COLUMN_DESCRIPTION}
  \item{\code{KitchenAbvGr}}{double COLUMN_DESCRIPTION}
  \item{\code{KitchenQual}}{character COLUMN_DESCRIPTION}
  \item{\code{TotRmsAbvGrd}}{double COLUMN_DESCRIPTION}
  \item{\code{Functional}}{character COLUMN_DESCRIPTION}
  \item{\code{Fireplaces}}{double COLUMN_DESCRIPTION}
  \item{\code{FireplaceQu}}{character COLUMN_DESCRIPTION}
  \item{\code{GarageType}}{character COLUMN_DESCRIPTION}
  \item{\code{GarageYrBlt}}{double COLUMN_DESCRIPTION}
  \item{\code{GarageFinish}}{character COLUMN_DESCRIPTION}
  \item{\code{GarageCars}}{double COLUMN_DESCRIPTION}
  \item{\code{GarageArea}}{double COLUMN_DESCRIPTION}
  \item{\code{GarageQual}}{character COLUMN_DESCRIPTION}
  \item{\code{GarageCond}}{character COLUMN_DESCRIPTION}
  \item{\code{PavedDrive}}{character COLUMN_DESCRIPTION}
  \item{\code{WoodDeckSf}}{double COLUMN_DESCRIPTION}
  \item{\code{OpenPorchSf}}{double COLUMN_DESCRIPTION}
  \item{\code{EnclosedPorch}}{double COLUMN_DESCRIPTION}
  \item{\code{X3SsnPorch}}{double COLUMN_DESCRIPTION}
  \item{\code{ScreenPorch}}{double COLUMN_DESCRIPTION}
  \item{\code{PoolArea}}{double COLUMN_DESCRIPTION}
  \item{\code{PoolQc}}{character COLUMN_DESCRIPTION}
  \item{\code{Fence}}{character COLUMN_DESCRIPTION}
  \item{\code{MiscFeature}}{character COLUMN_DESCRIPTION}
  \item{\code{MiscVal}}{double COLUMN_DESCRIPTION}
  \item{\code{MoSold}}{double COLUMN_DESCRIPTION}
  \item{\code{YrSold}}{double COLUMN_DESCRIPTION}
  \item{\code{SaleType}}{character COLUMN_DESCRIPTION}
  \item{\code{SaleCondition}}{character COLUMN_DESCRIPTION}
  \item{\code{SalePrice}}{double COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
HouseSalesData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
