% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or integer value of the year to add to the filename.}
}
\value{
A string being a modified filename for assignment.
}
\description{
This function generates a filename using a base string and adding a year 
passed via argument. The filename produced reflects the naming used in the 
fars data. Typically used as a private function.
}
\details{
Note this function does not consider the source data is in a directory and the 
likelihood the script is outside the directory. This is addressed by setting 
the working directory to be the data directory.
}
\examples{
\dontrun{
new_name <- make_filename(year = 2017)
new_name <- make_filename(2017)
}
}
