% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marylou}
\alias{marylou}
\title{marylou}
\format{
A data frame with 50 rows and 2 variables:
\describe{
  \item{\code{CustomerType}}{character:  adult or teenager}
  \item{\code{Spend}}{double amount of purchase when surveyed, in $} 
}
}
\source{
\url{adapted from exercise in Essentials of Business Statistics, 1st Edition, Sanjiv Jaggia and Alison Kelly , Copyright: 2014}
}
\usage{
marylou
}
\description{
age group and purchase amount for 50 customers of a coffee chain
}
\keyword{datasets}
