% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Función Mayores o menores.R
\name{mayores}
\alias{mayores}
\alias{menores}
\title{Valores extremos de una tabla}
\usage{
mayores(
  x,
  Q = 1,
  OrdenarPor = 2,
  ExtraerDe = 1,
  Vector = FALSE,
  Y = " y ",
  na.rm = FALSE
)

menores(
  x,
  Q = 1,
  OrdenarPor = 2,
  ExtraerDe = 1,
  Vector = FALSE,
  Y = " y ",
  na.rm = FALSE
)
}
\arguments{
\item{x}{Tabla, vector con nombres o data.frame.}

\item{Q}{Vector que indica qué datos se extraerán de la lista.}

\item{OrdenarPor}{Número que indica por qué columna se ordenará la tabla.}

\item{ExtraerDe}{Número que indica de qué columna se extraerá la lista.}

\item{Vector}{Valor lógico (TRUE o FALSE) que indica si se extraerá la lista
en un vector o como una cadena de texto en español.}

\item{Y}{Separador de la lista en caso de que sean más de dos elementos y que
el argumento 'Vector' sea FALSE.}

\item{na.rm}{Valor lógico que indica si se ignora los NA en la columna de
OrdenarPor.}
}
\value{
Un vector con los nombres de los valores extremos o una cadena de
texto con nombres los mismos valores extremos separados por comas y separando
con 'y' los últimos dos términos de forma que se pueda insertar dentro de un
documento de Word mediante un archivo rmarkdown en español.
}
\description{
Extrae los nombres de una tabla de acuerdo a los datos más altos o más bajos.
}
\details{
El argumento 'x' puede tener cualquiera de los siguentes formatos:
\itemize{
\item Una tabla o un vector con sus datos identificados por nombre.
\item Un data.frame del que se puede elegir qué columna se usará para ordenar y
qué columna se usará para extraer la lista de los valores extremos.
}

Se puede modificar el separador Y=' y ' por Y=' and ' para cambiar el idioma.
}
\examples{
# Ejemplo usando una tabla o un vector con nombres

## Creación de los datos  
x<-c(5,8,9,1,2,3)
x<-table(sample.int(6,60,replace = TRUE))

## Ejemplos
names(x)<-LETTERS[1:6]

Mayores(x,1:2) #Arroja las dos letras con más casos
Mayores(x,1:3,Vector = TRUE) #Arroja un vector de las tres letras con más casos

# Ejemplo usando un data.frame

x<-data.frame(Variable=LETTERS[1:6],F1=c(78,32,95,22,47,12),F2=c(1,7,2,6,9,3))

## Ejemplos

mayores(x) # Extrae de la primera columna el valor más grande ordenado por la
segunda columna.
menores(x,2) #Extrae el segundo valor más pequeño. 
mayores(x,1:4,OrdenarPor = 2,ExtraerDe = 3, Y=" and ") #Extrae de la tercer columna los 
cuatro valores más grandes ordenados por la segunda columna y en inglés.
}
