% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mileage}
\alias{mileage}
\title{mileage}
\format{
A data frame with 84 rows and 9 variables:
\describe{
  \item{\code{model}}{character:  car model}
  \item{\code{eng_size}}{double:  engine size, in liters}
  \item{\code{cylinders}}{double:  number of cylinders in engine}
  \item{\code{msrp}}{double:  suggested retail price of car, USD}
  \item{\code{city_mpg}}{double:  average miles per gallon, city driving}
  \item{\code{highway_mpg}}{double: average miles per gallon, highway driving}
  \item{\code{weight}}{double:  weight, pounds}
  \item{\code{type}}{character:  type of car}
  \item{\code{country}}{character:  country where manufacturer is based} 
}
}
\source{
{Sharpe, De Veaux, Velleman:  Business Statistics, A First Course, 2e, online chapters}
}
\usage{
mileage
}
\description{
fuel efficiency information for many types of cars
}
\keyword{datasets}
