% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varias funciones.R
\name{Modas}
\alias{Modas}
\title{Vector de modas}
\usage{
Modas(x)
}
\arguments{
\item{x}{Un vector con datos de cualquier tipo.}
}
\value{
Un vector con los valores que más se repiten.
}
\description{
Función genérica para obtener los valores que más se repiten.
}
\examples{

x <- c(rep(1,3),rep(2,4),rep(3,2),rep(4,1),rep(5,4),6)

Modas(x)
}
