% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{mortality}
\format{A data frame with 2021 rows and 2 variables:
\describe{
  \item{\code{ID}}{integer policy holder ID}
  \item{\code{Claimed}}{integer 1 if made claim on term life insturance; 0 otherwise}
}}
\usage{
mortality
}
\description{
Data on whether a term-life policy was claimed during the term.  Made-up for lab on one-sample h-tests for proportion.
}
\keyword{datasets}
