% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moviedata}
\alias{moviedata}
\title{moviedata}
\format{A data frame with 487 rows and 2 variables:
\describe{
  \item{\code{AgeGroup}}{character: age group of viewer}
  \item{\code{SeenBefore}}{character: had viewer seen before?} 
}}
\usage{
moviedata
}
\description{
made-up data for HW about movie attendance
}
\keyword{datasets}
