% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpgdata}
\alias{mpgdata}
\title{mpgdata}
\format{
A data frame with 84 rows and 9 variables:
\describe{
  \item{\code{Model}}{character:  model of car}
  \item{\code{Eng Size}}{double:  liters}
  \item{\code{Cylinders}}{double:  number of engine cylinders}
  \item{\code{MSRP}}{double:  suggested retail price, in USD}
  \item{\code{City Mpg}}{double:  city miles per gallon}
  \item{\code{Highway Mpg}}{double:  highway miles per gallon}
  \item{\code{Weight}}{double:  weight in pounds}
  \item{\code{Type}}{character:  car class}
  \item{\code{Country}}{character:  country of manufacturer} 
}
}
\source{
\url{Sharpe, DeVeaux, Velleman electronic chapters}
}
\usage{
mpgdata
}
\description{
auto fuel economy and other data
}
\keyword{datasets}
