% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varias funciones.R
\name{numerico}
\alias{numerico}
\alias{numerico2}
\title{Convertir a numérico para data.frames.}
\usage{
numerico(x, Columnas)

numerico2(x, Columnas)
}
\arguments{
\item{x}{Vector o data.frame a procesar}

\item{Columnas}{Columnas que se han de procesar. Aplica sólo si es data.frame.}
}
\value{
Un data.frame o un vector
}
\description{
Convierte un data.frame o un vector a valores numéricos.
}
\details{
Esta función es una cápsula para la función as.numeric que controla a qué
columnas se aplica la conversión, introduciendo NA en los datos que no se
pueden convertir. Cuando se le aplica a los factores, primero los convierte a
caracter y luego a numérico para respetar el valor observado.

La diferencia entre las dos funciones es que \emph{numerico} convierte las
columnas especificadas del dataframe y lo arroja completo, mientras que
\emph{numerico2} sólo arroja las columnas convertidas. Cuando se usa con vectores
el resultado difiere de as.numeric en que convierte los factores en el valor
observado y no en el número subyacente y en caso de que sean letras los
convierte en NA.
}
\examples{

Crear data.frame modelo
W<-data.frame( Variable=LETTERS[1:6],
               F1=factor(c(2,3,5,3,2,2)),
               F2=c(1,7,2,6,9,3),
               F3=c("Juan","Pedro","María","Fernanda","José","Pía"),
               F4=factor(c("H","H","M","M","H","M")))

#Uso:

#Genera el mismo data.frame con datos numéricos. Si hay columnas 
de caracteres introduce NA:
numerico(W) 

# Ambos ejemplos hacen lo mismo: Convierten las columnas 2:3 y las incorporan
al data.frame original 
W[2:3]<-numerico2(W,2:3)
W<-numerico(W,2:3)

# Si es un vector, genera un vector:
numerico(W$F1)
}
