% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PearsonClassSize}
\alias{PearsonClassSize}
\title{PearsonClassSize}
\format{A data frame with 300 rows and 6 variables:
\describe{
  \item{\code{campus}}{character:  which campus}
  \item{\code{class_size}}{double:  number of students in class}
  \item{\code{average_final_grade}}{double:  average final numeric grade}
  \item{\code{number_of_f_s}}{double:  how many Fs given in the class}
  \item{\code{average_g_p_a}}{double:  average GPA of students in class}
  \item{\code{successful_or_unsuccessful}}{character:} 
}}
\source{
{Pearson Publishing}
}
\usage{
PearsonClassSize
}
\description{
demogaraphic information about class sizes and performance
}
\keyword{datasets}
