% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PearsonCrackerBarrel}
\alias{PearsonCrackerBarrel}
\title{PearsonCrackerBarrel}
\format{
A data frame with 150 rows and 7 variables:
\describe{
  \item{\code{restaurant_number}}{double:  restaurant ID}
  \item{\code{geographic_region}}{character:  geographic region}
  \item{\code{annual_revenue}}{double:  annual revenue in $}
  \item{\code{average_cost_of_gasoline}}{double:  average cost of gasoline near restaurant}
  \item{\code{miles_from_interstate}}{double:  distance from nearest interstate, in miles}
  \item{\code{square_feet}}{double:  size of restaurant in square feet}
  \item{\code{annual_increase_in_revenue}}{character:  did revenue increase this year over the previous year, coded as y or n.} 
}
}
\source{
{Pearson Publishing}
}
\usage{
PearsonCrackerBarrel
}
\description{
information about various Cracker Barrel restaurants
}
\keyword{datasets}
