% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PearsonDealerships}
\alias{PearsonDealerships}
\title{PearsonDealerships}
\format{
A data frame with 1000 rows and 7 variables:
\describe{
  \item{\code{dealership_number}}{double:  dealership ID}
  \item{\code{state}}{character:  state dealership is located in}
  \item{\code{median_salary}}{double:  median salary of employees in $}
  \item{\code{annual_sales}}{double:  annual sales of dealership, in $}
  \item{\code{number_of_vehicles_sold}}{double:  number of vehicles sold in a period}
  \item{\code{square_feet}}{double:  square feet of dealership}
  \item{\code{quality_award_winner}}{character:  did they win a quality award, y or n} 
}
}
\source{
{Pearson Publishing}
}
\usage{
PearsonDealerships
}
\description{
information about car dealerships from Pearson
}
\keyword{datasets}
