% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PearsonIQ}
\alias{PearsonIQ}
\title{PearsonIQ}
\format{A data frame with 500 rows and 7 variables:
\describe{
  \item{\code{volunteer_number}}{double:  ID of respondent}
  \item{\code{gender}}{character:  gender}
  \item{\code{iq}}{double:  IQ score}
  \item{\code{annual_income}}{double:  annual income, in dollars}
  \item{\code{pre_test_score}}{double:  pre-test score}
  \item{\code{lifetime_savings}}{double:  lifetime savings, in $}
  \item{\code{gifted}}{character:  identified as gifted recorded as y or n} 
}}
\source{
\url{Pearson Publishing}
}
\usage{
PearsonIQ
}
\description{
IQ and demographic data for 500 respondents
}
\keyword{datasets}
