% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varias funciones.R
\name{permutacion}
\alias{permutacion}
\title{Permutación}
\usage{
permutacion(x, y)
}
\arguments{
\item{x}{Cantidad de elementos en el universo}

\item{y}{Cantidad de elementos a ordenar (sin remplazo)}
}
\value{
Un valor numérico del número de combinaciones
}
\description{
Función genérica para realizar una permutación de 'x' tomando 'y' (recuerden
que en la permutación el órden sí importa)
}
\examples{

permutación(5,3)
}
