% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RectanglesData}
\alias{RectanglesData}
\title{Rectangles Data}
\format{A data frame with 100 rows and 3 variables:
\describe{
  \item{\code{Number}}{integer COLUMN_DESCRIPTION}
  \item{\code{Actual Sizes}}{integer COLUMN_DESCRIPTION}
  \item{\code{OddEven}}{character COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
RectanglesData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
