% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shopgender}
\alias{shopgender}
\title{shopgender}
\format{
A data frame with 1085 rows and 2 variables:
\describe{
  \item{\code{Gender}}{character Male or Female}
  \item{\code{LikesToShop}}{character yes or no} 
}
}
\source{
\url{adapted from exercise in Business Statistics - A First Course, 6th Edition, Levine, Krehbiel, and Berenson, 2012}
}
\usage{
shopgender
}
\description{
simulated data to match USA Today poll about gender and shopping
}
\keyword{datasets}
