% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNSdata}
\alias{SNSdata}
\title{snsdata}
\format{
A data frame with 609 rows and 4 variables:
\describe{
  \item{\code{gender}}{character gender (male or female)}
  \item{\code{age}}{integer age}
  \item{\code{primary_sns}}{character primary social network used by tenn}
  \item{\code{how_often_visit}}{character how often do they visit their primary social network}
}
}
\source{
\url{http://www.pewinternet.org/dataset/september-2012-teens-and-online-privacy/}
}
\usage{
SNSdata
}
\description{
Excerpt from Pew Research Study on Teens, Social Media, and Privacy from May 21, 2013.  Survey of 12 - 17 year olds; only those teens who used at least one social network site were selected for this excerpt.
}
\keyword{datasets}
