% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpeedData}
\alias{SpeedData}
\title{SpeedData}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{\code{TYPE}}{character email interface application}
  \item{\code{SPEED}}{double email opening speed in seconds} 
}
}
\usage{
SpeedData
}
\description{
email download speeds
}
\keyword{datasets}
