% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractForm.R
\name{subtractForm}
\alias{subtractForm}
\title{subract linear fit to evaluate profile}
\usage{
subtractForm(travel, total, doPlot = TRUE)
}
\arguments{
\item{travel}{}

\item{total}{}

\item{doPlot}{}
}
\value{
list of corrected and uncorrected x and y
}
\description{
subract linear fit to evaluate profile
}
