% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varias funciones.R
\name{sustituye}
\alias{sustituye}
\title{Sustituir valores en un DF}
\usage{
sustituye(Previo, Nuevo, Donde)
}
\arguments{
\item{Previo}{Caracter que se va a cambiar.}

\item{Nuevo}{Caracter que se va a incluir}

\item{Donde}{Data.frame o vector en el que se hará la sustitución}
}
\value{
Un vector o un data.frame
}
\description{
Sustituye un texto en cualquier parte de un data.frame por otro texto.
}
\details{
OJO: Convierte todas las columnas a caracteres.
}
\examples{

# Crear datos modelo
x<-data.frame(Variable=LETTERS[1:6],F1=factor(c(2,3,5,32,6,2)),F2=c(1,7,2,6,9,3))

# Ejemplo 
sustituye(2,"R",x)
}
