% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UltraGreenData}
\alias{UltraGreenData}
\title{UltraGreenData}
\format{A data frame with 25 rows and 2 variables:
\describe{
  \item{\code{carID}}{integer:  ID}
  \item{\code{mpg}}{integer:  miles per gallon} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
UltraGreenData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
