% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectoraLista.R
\name{VectoraLista}
\alias{VectoraLista}
\title{Vector a lista de datos.}
\usage{
VectoraLista(x, Y = " y ")
}
\arguments{
\item{x}{Vector}

\item{Y}{Caracter que separa los últimos dos datos.}
}
\value{
Un vector con los valores mayores de otra columna o una cadena de
texto en español.
}
\description{
Convierte los datos de un vector en una lista de datos en español.
}
\details{
La función produce una cadena de texto con cada elemento separado por una
coma y los últimos dos por la palabra "y". Alternativamente se puede
sustituir la palabra y por otra para escribir en otro idioma. El parámetro
'Y' debe incluir los espacios entre el número y la palabra, también se puede
agregar la coma como parte del separador.
}
\examples{

# Crear vector modelo
x<-LETTERS[1:10]

# Ejemplo

VectoraLista(x)
VectoraLista(x,Y=" and ") # Cambia la lista a inglés
}
