% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VendorSelectionData}
\alias{VendorSelectionData}
\title{DATASET_TITLE}
\format{A data frame with 454 rows and 3 variables:
\describe{
  \item{\code{item}}{character:  item ID}
  \item{\code{vendorA}}{character:  value appraised by Vendor A}
  \item{\code{vendorB}}{character:  value appraised by Vendor B} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
VendorSelectionData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
