% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph Dataset}
\format{
A data frame with 12 rows and 3 columns:
\describe{
  \item{v1}{Starting node (integer)}
  \item{v2}{Ending node (integer)}
  \item{w}{Weight of the edge between nodes (integer)}
}
}
\source{
Wikipedia example graph
}
\usage{
wiki_graph
}
\description{
A dataset representing a weighted graph with six nodes and edges with weights.
}
\examples{
data(wiki_graph)
wiki_graph
}
\keyword{datasets}
