% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varias funciones.R
\name{z.test}
\alias{z.test}
\title{Título}
\usage{
z.test(x, p, na = TRUE)
}
\arguments{
\item{na.rm}{Valor lógico que indica si se ignora los NA en la columna de}
}
\value{
Un vector con los nombres de los valores extremos o una cadena de
}
\description{
Descripción breve.
}
\details{
Detalles
}
