% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZagatData}
\alias{ZagatData}
\title{ZagatData}
\format{
A data frame with 100 rows and 6 variables:
\describe{
  \item{\code{location}}{character: restaurant location, suburban or city}
  \item{\code{food}}{integer: food rating on 30 point scale}
  \item{\code{decor}}{integer: decor rating on 30 point scale}
  \item{\code{service}}{integer: service rating on 30 point scale}
  \item{\code{totalrating}}{integer: sum of food, decor, and service ratings}
  \item{\code{avgprice}}{integer: average menu price per person} 
}
}
\source{
{from Zagat}
}
\usage{
ZagatData
}
\description{
Zagat survey ratings with average menu prices and locations
}
\keyword{datasets}
