% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AirlineData}
\alias{AirlineData}
\title{AirlineData}
\format{A data frame with 50 rows and 4 variables:
\describe{
  \item{\code{airline}}{character: airline}
  \item{\code{status}}{character: delay status}
  \item{\code{minutes}}{integer: actual number of minutes deviation from schedule, negative indicates minutes early, positive indicates minutes late)}
}}
\source{
{}
}
\usage{
AirlineData
}
\description{
On-time performance of airline arrivals
}
\keyword{datasets}
