% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarInsData}
\alias{CarInsData}
\title{CarInsData}
\format{A data frame with 500 rows and 31 variables:
\describe{
  \item{\code{Obs}}{double COLUMN_DESCRIPTION}
  \item{\code{ID}}{double COLUMN_DESCRIPTION}
  \item{\code{Crash}}{double COLUMN_DESCRIPTION}
  \item{\code{CrashCost}}{double COLUMN_DESCRIPTION}
  \item{\code{DriverAge}}{double COLUMN_DESCRIPTION}
  \item{\code{LevelEducation}}{character COLUMN_DESCRIPTION}
  \item{\code{YearsJob}}{double COLUMN_DESCRIPTION}
  \item{\code{Income}}{double COLUMN_DESCRIPTION}
  \item{\code{SexofDriver}}{character COLUMN_DESCRIPTION}
  \item{\code{TypeofJob}}{character COLUMN_DESCRIPTION}
  \item{\code{TimeCommute}}{double COLUMN_DESCRIPTION}
  \item{\code{PointsonRecord}}{double COLUMN_DESCRIPTION}
  \item{\code{LicenseRevoked}}{character COLUMN_DESCRIPTION}
  \item{\code{ValueofHome}}{double COLUMN_DESCRIPTION}
  \item{\code{MaritalStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{SingleParent}}{character COLUMN_DESCRIPTION}
  \item{\code{NumberKidsHome}}{double COLUMN_DESCRIPTION}
  \item{\code{NumberKidsDrive}}{double COLUMN_DESCRIPTION}
  \item{\code{UrbanorRural}}{character COLUMN_DESCRIPTION}
  \item{\code{AgeofCar}}{double COLUMN_DESCRIPTION}
  \item{\code{TypeofCar}}{character COLUMN_DESCRIPTION}
  \item{\code{TypeCarUse}}{character COLUMN_DESCRIPTION}
  \item{\code{BlueBookValue}}{double COLUMN_DESCRIPTION}
  \item{\code{RedCar}}{character COLUMN_DESCRIPTION}
  \item{\code{TimeasCustomer}}{double COLUMN_DESCRIPTION}
  \item{\code{NumberClaims5yr}}{double COLUMN_DESCRIPTION}
  \item{\code{Total5YrPayout}}{double COLUMN_DESCRIPTION}
  \item{\code{NotWorking}}{character COLUMN_DESCRIPTION}
  \item{\code{NewCar}}{character COLUMN_DESCRIPTION}
  \item{\code{GoodDriver}}{character COLUMN_DESCRIPTION}
  \item{\code{OwnHome}}{character COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
CarInsData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
