% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoffeeSaleData}
\alias{CoffeeSaleData}
\title{CoffeeSaleData}
\format{A data frame with 50 rows and 1 variables:
\describe{
  \item{\code{sales}}{double: customer level sales in US dollars} 
}}
\source{
{}
}
\usage{
CoffeeSaleData
}
\description{
Sample of individual customer coffee sales
}
\keyword{datasets}
