% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoffeeTypeData}
\alias{CoffeeTypeData}
\title{CoffeeTypeData}
\format{A data frame with 50 rows and 1 variables:
\describe{
  \item{\code{type}}{character:  was customer an adult woman or a teen girl} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
CoffeeTypeData
}
\description{
classification of a random sample of 50 customers from a coffee shop
}
\keyword{datasets}
