% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CreditCardData}
\alias{CreditCardData}
\title{CreditCardData}
\format{A data frame with 401 rows and 11 variables:
\describe{
  \item{\code{id}}{integer: identifier}
  \item{\code{income}}{double: in thousands of US dollars}
  \item{\code{limit}}{integer: credit limit in US dollars}
  \item{\code{rating}}{integer: internal bank credit rating, a number between 300 and 850}
  \item{\code{cards}}{integer: number of credit cards}
  \item{\code{age}}{integer: in years}
  \item{\code{education}}{integer: years of education}
  \item{\code{gender}}{character: male or female}
  \item{\code{student}}{character: customer student indicator, yes or no}
  \item{\code{married}}{character: customer marital status indicator, yes or no}
  \item{\code{balance}}{integer: average credit card debt in US dollars} 
}}
\source{
{}
}
\usage{
CreditCardData
}
\description{
Data set of credit card customers
}
\keyword{datasets}
