% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CustValue}
\alias{CustValue}
\title{CustValue}
\format{
A data frame with 9134 rows and 24 variables:
\describe{
  \item{\code{customer}}{character Customer ID}
  \item{\code{state}}{character U.S. state of residence}
  \item{\code{customer_lifetime_value}}{double value of customer to company in USD}
  \item{\code{response}}{character unsure what this is:  if responded to marketing?}
  \item{\code{coverage}}{character type of auto insurance policy}
  \item{\code{education}}{character level of education}
  \item{\code{effective_to_date}}{character }
  \item{\code{employment_status}}{character employment status}
  \item{\code{gender}}{character gender}
  \item{\code{income}}{double income, USD}
  \item{\code{location_code}}{character type of location of policyholder residence}
  \item{\code{marital_status}}{character marital status}
  \item{\code{monthly_premium_auto}}{double monthly cost of auto premium}
  \item{\code{months_since_last_claim}}{double number of months since last claim}
  \item{\code{months_since_policy_inception}}{double months since became a policyholder}
  \item{\code{number_of_open_complaints}}{double number of open complaints (unsure made by whom about who)}
  \item{\code{number_of_policies}}{double number of policies (other insurance policies by same company)}
  \item{\code{policy_type}}{character type of auto policy}
  \item{\code{policy}}{character further breakdown of type of auto policy}
  \item{\code{renew_offer_type}}{character what type of renewal offer were they given?}
  \item{\code{sales_channel}}{character sales channel}
  \item{\code{total_claim_amount}}{double total amount of claims made by policyholder?? seems low}
  \item{\code{vehicle_class}}{character type of vehicle insured}
  \item{\code{vehicle_size}}{character size of insured vehicle} 
}
}
\source{
\url{https://www.kaggle.com/pankajjsh06/ibm-watson-marketing-customer-value-data}
}
\usage{
CustValue
}
\description{
Simulated data about auto insurance company and value of customer, by IBM
}
\keyword{datasets}
