% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decindep}
\alias{decindep}
\title{decindep}
\format{
A data frame with 1323 rows and 3 variables:
\describe{
  \item{\code{Word}}{character: word}
  \item{\code{Length}}{double:  number of letters in word}
  \item{\code{Has_an_e}}{logical:  TRUE if word contains an e} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
decindep
}
\description{
words in the U.S. Declaration of Independence
}
\keyword{datasets}
