% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EmailAccessData}
\alias{EmailAccessData}
\title{EmailAccessData}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{\code{emailv1}}{double:  time}
  \item{\code{emailv2}}{double:  time} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
EmailAccessData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
