% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, years)
}
\arguments{
\item{state.num}{The number of the state to filter the observations for}

\item{years}{The year(s) of the accident data to use}
}
\value{
A map object containing plotted points for the location of each observation in a specified state, with a LONGITUD value over 900, and a LATITUDE value over 90, during a specified year
}
\description{
A function will plot the location of the observations with a LONGITUD value over 900, and a LATITUDE value over 90, in a specified state, during a specified year, on a map
}
\note{
If the value provided for the state.num is not of type numeric, and the value provided for the year parameter is not of type integer, the function will attempt to coerce it into a values of types numeric and integer, and if it cannot do so, an error fill be thrown
}
\examples{
alabama_accidents <- fars_map_state(1, 2015)

}
