% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The CSV file to turn into an object of the tbl_df class}
}
\value{
The tbl_df dataframe object
}
\description{
A function that converts the accident data to an object of the tbl_df class, a subclass of data.frame
}
\note{
If the data cannot be read, the program will attempt to read accident_xxx.CSV from the root dir. If neither can be read, an error will be thrown
}
\examples{
csvdata <- fars_read("accident_2015")

}
