% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{The year(s) of the accident CSV data file(s) to use}
}
\value{
The tbl_df dataframe object, containing the year and month of the found observations
}
\description{
A function that will read the accident CSV file(s) for given year(s), and return the year and month of the found observations
}
\note{
If the value provided for the year parameter is not of type integer, the function will attempt to coerce it into a value of type integer, and if it cannot do so, an error fill be thrown
}
\examples{
accident_ym <- fars_read_years(c(2014, 2015))

}
