% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{The year(s) of the accident data to use}
}
\value{
The tbl_df dataframe object, containing the summary
}
\description{
A function that will summarize the number of observations for every month, in the accident data, by years
}
\note{
If the value provided for the year parameter is not of type integer, the function will attempt to coerce it into a value of type integer, and if it cannot do so, an error fill be thrown
}
\examples{
accidentsummary <- fars_summarize_years(c(2014, 2015))

}
