% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HeartData}
\alias{HeartData}
\title{HeartData}
\format{
A data frame with 303 rows and 14 variables:
\describe{
  \item{\code{age}}{double }
  \item{\code{sex}}{double }
  \item{\code{cp}}{double }
  \item{\code{trestbps}}{double }
  \item{\code{chol}}{double }
  \item{\code{fbs}}{double }
  \item{\code{restecg}}{double }
  \item{\code{thalach}}{double }
  \item{\code{exang}}{double }
  \item{\code{oldpeak}}{double }
  \item{\code{slope}}{double }
  \item{\code{ca}}{double }
  \item{\code{thal}}{double }
  \item{\code{target}}{double } 
}
}
\usage{
HeartData
}
\description{
variables about heart health events
}
\keyword{datasets}
