% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HouseSalesData}
\alias{HouseSalesData}
\title{HouseSalesData}
\format{A data frame with 1460 rows and 80 variables:
\describe{
  \item{\code{Id}}{double Identifier}
  \item{\code{MsZoning}}{character zoning type}
  \item{\code{LotFrontage}}{double linear feet of property frontage to street}
  \item{\code{LotArea}}{double area in square feet of plot}
  \item{\code{Street}}{character type of street}
  \item{\code{Alley}}{character indicator for alley}
  \item{\code{LotShape}}{character descriptor for shape of land plot}
  \item{\code{LandContour}}{character contour of land}
  \item{\code{Utilities}}{character utility type to property}
  \item{\code{LotConfig}}{character configuration of plot}
  \item{\code{LandSlope}}{character how sloped is the plot}
  \item{\code{Neighborhood}}{character neighborhood}
  \item{\code{Condition1}}{character }
  \item{\code{Condition2}}{character }
  \item{\code{BldgType}}{character }
  \item{\code{HouseStyle}}{character }
  \item{\code{OverallQual}}{double quality rating, 1 – 10 scale}
  \item{\code{OverallCond}}{double condition rating, 1 – 10 scale}
  \item{\code{YearBuilt}}{double }
  \item{\code{YearRemodAdd}}{double }
  \item{\code{RoofStyle}}{character type of roof on property}
  \item{\code{RoofMatl}}{character material used in roof}
  \item{\code{Exterior1St}}{character }
  \item{\code{Exterior2Nd}}{character }
  \item{\code{MasVnrType}}{character }
  \item{\code{MasVnrArea}}{double COLUMN }
  \item{\code{ExterQual}}{character }
  \item{\code{ExterCond}}{character }
  \item{\code{Foundation}}{character }
  \item{\code{BsmtQual}}{character }
  \item{\code{BsmtCond}}{character }
  \item{\code{BsmtExposure}}{character }
  \item{\code{BsmtFinType1}}{character }
  \item{\code{BsmtFinSf1}}{double }
  \item{\code{BsmtFinType2}}{character }
  \item{\code{BsmtFinSf2}}{double }
  \item{\code{BsmtUnfSf}}{double }
  \item{\code{TotalBsmtSf}}{double }
  \item{\code{Heating}}{character }
  \item{\code{HeatingQc}}{character }
  \item{\code{CentralAir}}{character }
  \item{\code{Electrical}}{character }
  \item{\code{X1StFlrSf}}{double }
  \item{\code{X2NdFlrSf}}{double }
  \item{\code{LowQualFinSf}}{double }
  \item{\code{GrLivArea}}{double }
  \item{\code{BsmtFullBath}}{double }
  \item{\code{BsmtHalfBath}}{double }
  \item{\code{FullBath}}{double }
  \item{\code{HalfBath}}{double }
  \item{\code{BedroomAbvGr}}{double }
  \item{\code{KitchenAbvGr}}{double }
  \item{\code{KitchenQual}}{character }
  \item{\code{TotRmsAbvGrd}}{double }
  \item{\code{Functional}}{character }
  \item{\code{Fireplaces}}{double }
  \item{\code{FireplaceQu}}{character }
  \item{\code{GarageType}}{character }
  \item{\code{GarageYrBlt}}{double }
  \item{\code{GarageFinish}}{character }
  \item{\code{GarageCars}}{double }
  \item{\code{GarageArea}}{double }
  \item{\code{GarageQual}}{character }
  \item{\code{GarageCond}}{character }
  \item{\code{PavedDrive}}{character }
  \item{\code{WoodDeckSf}}{double }
  \item{\code{OpenPorchSf}}{double }
  \item{\code{EnclosedPorch}}{double }
  \item{\code{X3SsnPorch}}{double }
  \item{\code{ScreenPorch}}{double }
  \item{\code{PoolArea}}{double }
  \item{\code{PoolQc}}{character }
  \item{\code{Fence}}{character }
  \item{\code{MiscFeature}}{character }
  \item{\code{MiscVal}}{double }
  \item{\code{MoSold}}{double }
  \item{\code{YrSold}}{double }
  \item{\code{SaleType}}{character }
  \item{\code{SaleCondition}}{character }
  \item{\code{SalePrice}}{double } 
}}
\usage{
HouseSalesData
}
\description{
Collection of variables that describe house sales in one community.
}
\keyword{datasets}
