% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PopcornData}
\alias{PopcornData}
\title{PopcornData}
\format{A data frame with 134 rows and 5 variables:
\describe{
  \item{\code{id}}{character:  identifier}
  \item{\code{rank}}{character:  scout rank}
  \item{\code{grade}}{integer:  scout's grade in school}
  \item{\code{sale}}{double:  value of popcorn sold by scout, in dollars}
  \item{\code{year}}{integer: year of sale} 
}}
\source{
{data collected from Cub Scout Pack 658 by Linda Boardman Liu}
}
\usage{
PopcornData
}
\description{
Popcorn sales by Cub Scouts in 2017 and 2018
}
\keyword{datasets}
