% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RectanglesData}
\alias{RectanglesData}
\title{RectanglesData}
\format{A data frame with 100 rows and 3 variables:
\describe{
  \item{\code{Number}}{integer: identifier}
  \item{\code{Actual Sizes}}{integer: number of rectangles}
  \item{\code{OddEven}}{character: categorical indicating odd or even number of rectangles} 
}}
\source{
{}
}
\usage{
RectanglesData
}
\description{
Actual sizes of rectangles exercise
}
\keyword{datasets}
