% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpeedData}
\alias{SpeedData}
\title{SpeedData}
\format{A data frame with 30 rows and 2 variables:
\describe{
  \item{\code{TYPE}}{character COLUMN_DESCRIPTION}
  \item{\code{SPEED}}{double COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
SpeedData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
