% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UltraGreenData}
\alias{UltraGreenData}
\title{UltraGreenData}
\format{
A data frame with 25 rows and 2 variables:
\describe{
  \item{\code{carID}}{integer: identifier}
  \item{\code{mpg}}{integer: average miles per gallon} 
}
}
\source{
{}
}
\usage{
UltraGreenData
}
\description{
Ultra green car mileage per gallon
}
\keyword{datasets}
