% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VendorSelectionData}
\alias{VendorSelectionData}
\title{VendorSelectionData}
\format{A data frame with 454 rows and 3 variables:
\describe{
  \item{\code{item}}{character: identifier}
  \item{\code{vendorA}}{double: Vendor A replacement value quote}
  \item{\code{vendorB}}{double: Vendor B replacement value quote} 
}}
\source{
{collected from insurance company}
}
\usage{
VendorSelectionData
}
\description{
insurance replacement values for a collection of consumer electronics
}
\keyword{datasets}
