% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZagatData}
\alias{ZagatData}
\title{DATASET_TITLE}
\format{A data frame with 100 rows and 6 variables:
\describe{
  \item{\code{location}}{character restaurant location:  Suburban or City}
  \item{\code{food}}{integer:  food rating, out of 30}
  \item{\code{decor}}{integer:  decor rating, out of 30}
  \item{\code{service}}{integer:  service rating, out of 30}
  \item{\code{totalrating}}{integer:  sum of food, decor, and service ratings}
  \item{\code{avgprice}}{integer:  average price of a meal at the restaurant} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ZagatData
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
